[FIX] sale_variant_configurator: Populate product description in customer language
As the model `sale.order.line` doesn't have `partner_id` field, the
description is not populated in the customer language, as the automatism
put in product_configurator makes use of this trick.

For not adding such field that can conflict with other parts, let's
overwrite the methods that populate the description and pass upstream
the customer language in the context.

TT33029